KT-07: CRUD Question metadata ownership, R Roles, R Ownerships, R Users by role

+++ src/main/java/com/wojcka/exammanager/controllers/QuestionMetadataController.java
@@ -44,7 +44,7 @@ public class QuestionMetadataController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Void> createQuestionMetadata(@PathVariable("id") @Min(1) Long id, @Valid @RequestBody QuestionMetadata request) {
+    public ResponseEntity<Void> createQuestionMetadata(@PathVariable("id") @Min(1) Integer id, @Valid @RequestBody QuestionMetadata request) {
         request.setId(id);
 
         log.info(ObjectToJson.toJson(request));
